Implemented: Shifted getProductIdentificationValue function from utils/product/index.ts to utils/index.ts
