+++ src/utils/index.ts
@@ -1,7 +1,6 @@
 import { toastController } from '@ionic/vue';
 import { translate } from '@/i18n'
 import { Plugins } from '@capacitor/core';
-import productHelpers from './product'
 import { DateTime } from "luxon";
 
 // TODO Use separate files for specific utilities
@@ -39,4 +38,25 @@ const handleDateTimeInput = (dateTimeValue: any) => {
   return DateTime.fromISO(dateTime).toMillis()
 }
 
-export { handleDateTimeInput, showToast, hasError, copyToClipboard, productHelpers }
+const getProductIdentificationValue = (productIdentifier: string, product: any) => {
+
+  // handled this case as on page load initially the data is not available, so not to execute furthur code
+  // untill product are not available
+  if (!Object.keys(product).length) {
+    return;
+  }
+
+  let value = product[productIdentifier]
+
+  // considered that the goodIdentification will always have values in the format "productIdentifier/value" and there will be no entry like "productIdentifier/"
+  const identification = product['goodIdentifications'].find((identification: string) => identification.startsWith(productIdentifier + "/"))
+
+  if (identification) {
+    const goodIdentification = identification.split('/')
+    value = goodIdentification[1]
+  }
+
+  return value;
+}
+
+export { handleDateTimeInput, showToast, hasError, copyToClipboard, getProductIdentificationValue }
